PAP::Command::BlastP -Was setting the blastp_category tag to 'protein_category' instead of  $protein_category.
PAP::Command::DeleteFiles
-Make sure files exist before trying to unlink them. 


git-svn-id: svn+ssh://svn/srv/svn/gscpan/perl_modules/trunk@40741 127c0c0e-971a-0410-ad48-868c1a942d8d

+++ PAP/Command/DeleteFiles.pm
@@ -47,10 +47,12 @@ sub execute {
 
     foreach my $file (@files) {
 
-        unless(unlink($file)) {
+    	if (-e $file) {
+            unless(unlink($file)) {
 
-            die "failed to unlink '$file': $OS_ERROR";
+                die "failed to unlink '$file': $OS_ERROR";
 
+         	}	
         }
 
     }
